Refactor MetricReport to be a service
